+++ crates/ryo3-fnv/src/fnv1a.rs
@@ -213,7 +213,7 @@ impl PyFnv1a {
 impl std::fmt::Display for PyFnv1a {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let seed = self.finish().expect("no-way-jose");
-        write!(f, "fnv1a<{seed:x}>")
+        write!(f, "fnv1a<{seed:016x}>")
     }
 }