+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 
 ---
 
+## TBD
+
+- `ryo3-crc32fast`
+  - `crc32` class hasher implementation following the ry hasher conventions; I
+    need this soon and need it to be fast. The `crc32fast` crate is already used
+    by the `flate2` crate
+
 ## v0.0.100 [unreleased]
 
 - `ryo3-jiff`